requirements.txt
asgiref==3.5.1
Django==4.0.4
django-autoslug==1.9.8
django-ckeditor==6.4.0
django-js-asset==2.0.0
Pillow==9.1.0
sqlparse==0.4.2
